EPMEDU-728 - moved preview to the bottom
