+++ feature/more/src/main/java/com/epmedu/animeal/more/root/MoreScreenUi.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.epmedu.animeal.base.theme.AnimealTheme
-import com.epmedu.animeal.base.ui.AnimealShortButton
+import com.epmedu.animeal.foundation.topbar.AnimealShortButton
 import com.epmedu.animeal.foundation.topbar.TopBar
 import com.epmedu.animeal.more.root.ui.MoreOption
 import com.epmedu.animeal.more.screens
