+++ feature/splash/src/main/java/com/epmedu/animeal/splash/presentation/SplashScreenUI.kt
@@ -1,0 +1,64 @@
+package com.epmedu.animeal.splash.presentation
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.epmedu.animeal.base.theme.AnimealTheme
+import com.epmedu.animeal.foundation.spacer.HSpacer
+import com.epmedu.animeal.resources.R
+
+@Composable
+internal fun SplashScreenUI(
+    modifier: Modifier = Modifier,
+) {
+    Surface(
+        modifier = modifier
+            .fillMaxSize()
+            .background(
+                color = MaterialTheme.colors.background,
+            )
+    ) {
+        Column(
+            modifier = Modifier.fillMaxSize(),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center,
+        ) {
+            Image(
+                modifier = Modifier.width(110.dp),
+                contentScale = ContentScale.Crop,
+                painter = painterResource(
+                    id = R.drawable.ic_animeal_logo
+                ),
+                contentDescription = null
+            )
+            HSpacer(16.dp)
+            Text(
+                text = stringResource(id = R.string.app_name).uppercase(),
+                style = MaterialTheme.typography.h6,
+            )
+        }
+    }
+}
+
+@Suppress("UnusedPrivateMember")
+@Preview
+@Composable
+private fun SplashScreenUIPreview() {
+    AnimealTheme {
+        SplashScreenUI()
+    }
+}
